app/src/requirement.txt
loguru
fastapi
uvicorn
joblib
python-dotenv
